# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MyBank API",
    "description": "API per progetto universitario bancario MyBank sviluppato con Next.js",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/"
    }
  ],
...
  "paths": {
...
    "/transactions/withdrawal": {
      "post": {
        "tags": ["Transactions"],
        "summary": "Esegue un prelievo da un conto bancario",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["account_id", "amount"],
                "properties": {
                  "account_id": {
                    "type": "integer",
                    "example": 123,
                    "description": "Id del conto su cui effettuare il prelievo"
                  },
                  "amount": {
                    "type": "number",
                    "format": "float",
                    "example": 50.0,
                    "description": "Importo del prelievo"
                  },
                  "description": {
                    "type": "string",
                    "example": "Prelievo bancomat",
                    "description": "Descrizione opzionale del prelievo"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prelievo completato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "account_id": {
                          "type": "integer"
                        },
                        "transaction_type": {
                          "type": "string"
                        },
                        "amount": {
                          "type": "number",
                          "format": "float"
                        },
                        "balance_after": {
                          "type": "number",
                          "format": "float"
                        },
                        "description": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Errore di validazione input o saldo insufficiente"
          },
          "401": {
            "description": "Token mancante o non valido"
          },
          "500": {
            "description": "Errore nel processamento del prelievo"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}